Move radioStreams data array file to assets/data/ directory for improved folder structuring
