+++ app/assets/data/radioStreams.ts
@@ -9,7 +9,7 @@ export const playListData: Track[] = [
     title: 'Radar 107.5',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/radar107.5-albumart.png'),
+      require('../images/radar107.5-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8268/stream',
     isLiveStream: true,
@@ -19,7 +19,7 @@ export const playListData: Track[] = [
     title: 'Crystal 101.1',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/cristal101.1-albumart.png'),
+      require('../images/cristal101.1-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8274/stream',
     isLiveStream: true,
@@ -29,7 +29,7 @@ export const playListData: Track[] = [
     title: 'Radio Lobo Bajío 88.1 FM & 920 AM (Celaya)',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/lobobajio88.1-920am-albumart.png'),
+      require('../images/lobobajio88.1-920am-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8188/stream',
     isLiveStream: true,
@@ -39,7 +39,7 @@ export const playListData: Track[] = [
     title: 'Radar 88.9 FM (León)',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/radar88.9-albumart.png'),
+      require('../images/radar88.9-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8382/stream',
     isLiveStream: true,
@@ -49,7 +49,7 @@ export const playListData: Track[] = [
     title: 'El y Ella 103.7 FM (Celaya)',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/elyella103.7-albumart.png'),
+      require('../images/elyella103.7-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8272/stream',
     isLiveStream: true,
@@ -59,7 +59,7 @@ export const playListData: Track[] = [
     title: 'Stereo Cristal 96.7  (Celaya)',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/cristal96.7-albumart.png'),
+      require('../images/cristal96.7-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8264/stream',
     isLiveStream: true,
@@ -69,7 +69,7 @@ export const playListData: Track[] = [
     title: 'Radio Lobo 88.3 FM (San José Iturbide)',
     artist: 'Grupo Radar',
     artwork: Image.resolveAssetSource(
-      require('./assets/images/lobo88.3-albumart.png'),
+      require('../images/lobo88.3-albumart.png'),
     ).uri,
     url: 'https://radios.radiohd.com.mx/8270/stream',
     isLiveStream: true,
