+++ app/screens/MusicPlayer.tsx
@@ -6,7 +6,7 @@ import TrackPlayer, {
   Track,
   useTrackPlayerEvents,
 } from 'react-native-track-player';
-import {playListData} from '../radioStreams';
+import {playListData} from '../assets/data/radioStreams';
 import SongInfo from '../components/SongInfo';
 import ControlCenter from '../components/ControlCenter';
